nanodet-jittor/tools/test_pretrained_inference.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

"""
测试预训练权重推理
验证Jittor版本能否直接使用ImageNet预训练权重进行推理
"""

import os
import sys
import cv2
import numpy as np
import jittor as jt

# 添加项目路径
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from nanodet.model import build_model


def create_model():
    """创建NanoDet模型"""
    model_cfg = {
        'name': 'NanoDetPlus',
        'backbone': {
            'name': 'ShuffleNetV2',
            'model_size': '1.0x',
            'out_stages': [2, 3, 4],
            'activation': 'LeakyReLU',
            'pretrain': True  # 加载ImageNet预训练权重
        },
        'fpn': {
            'name': 'GhostPAN',
            'in_channels': [116, 232, 464],
            'out_channels': 96,
            'kernel_size': 5,
            'num_extra_level': 1,
            'use_depthwise': True,
            'activation': 'LeakyReLU'
        },
        'aux_head': {
            'name': 'SimpleConvHead',
            'num_classes': 20,
            'input_channel': 192,
            'feat_channels': 192,
            'stacked_convs': 4,
            'strides': [8, 16, 32, 64],
            'activation': 'LeakyReLU',
            'reg_max': 7
        },
        'head': {
            'name': 'NanoDetPlusHead',
            'num_classes': 20,
            'input_channel': 96,
            'feat_channels': 96,
            'stacked_convs': 2,
            'kernel_size': 5,
            'strides': [8, 16, 32, 64],
            'conv_type': 'DWConv',
            'norm_cfg': dict(type='BN'),
            'reg_max': 7,
            'activation': 'LeakyReLU',
            'loss': {
                'loss_qfl': {'beta': 2.0, 'loss_weight': 1.0},
                'loss_dfl': {'loss_weight': 0.25},
                'loss_bbox': {'loss_weight': 2.0}
            }
        },
        'detach_epoch': 10
    }
    
    return build_model(model_cfg)


def preprocess_image(img_path, input_size=(320, 320)):
    """预处理真实图像"""
    img = cv2.imread(img_path)
    if img is None:
        raise ValueError(f"Cannot load image: {img_path}")

    # 调整大小
    img = cv2.resize(img, input_size)

    # 归一化 (ImageNet标准)
    mean = np.array([103.53, 116.28, 123.675])
    std = np.array([57.375, 57.12, 58.395])

    img = img.astype(np.float32)
    img -= mean
    img /= std

    # 转换为CHW格式
    img = img.transpose(2, 0, 1)

    # 添加batch维度
    img = np.expand_dims(img, axis=0)

    return jt.array(img)


def test_pretrained_inference():
    """测试预训练权重推理"""
    print("=" * 60)
    print("测试预训练权重推理")
    print("=" * 60)
    
    # 设置CUDA
    if jt.has_cuda:
        jt.flags.use_cuda = 1
        print("✓ Using CUDA")
    
    # 创建模型
    print("创建模型...")
    model = create_model()
    model.eval()
    
    print(f"✓ 模型创建成功")
    print(f"  参数数量: {sum(p.numel() for p in model.parameters())/1e6:.2f}M")
    
    # 测试前向传播
    print("\n测试前向传播...")
    test_input = jt.randn(1, 3, 320, 320)
    
    with jt.no_grad():
        try:
            output = model(test_input)
            print(f"✓ 前向传播成功")
            print(f"  输入形状: {test_input.shape}")
            print(f"  输出形状: {output.shape}")
            print(f"  输出范围: [{output.min():.4f}, {output.max():.4f}]")
            
            # 检查输出是否合理
            if output.shape[0] == 1 and output.shape[1] > 1000:  # 应该有很多anchor点
                print("✓ 输出形状合理")
            else:
                print("⚠ 输出形状可能不正确")
                
        except Exception as e:
            print(f"✗ 前向传播失败: {e}")
            return False
    
    # 测试真实VOC图像
    test_images = [
        "data/VOCdevkit/VOC2007/JPEGImages/000001.jpg",
        "data/VOCdevkit/VOC2007/JPEGImages/000002.jpg",
        "data/VOCdevkit/VOC2007/JPEGImages/000003.jpg",
        "data/VOCdevkit/VOC2007/JPEGImages/000004.jpg",
        "data/VOCdevkit/VOC2007/JPEGImages/000005.jpg"
    ]

    successful_tests = 0
    for img_path in test_images:
        if os.path.exists(img_path):
            print(f"\n测试真实VOC图像: {img_path}")
            try:
                img_tensor = preprocess_image(img_path)
                with jt.no_grad():
                    output = model(img_tensor)
                print(f"✓ 图像推理成功: {output.shape}")
                print(f"  输出范围: [{output.min():.4f}, {output.max():.4f}]")
                successful_tests += 1
            except Exception as e:
                print(f"⚠ 图像推理失败: {e}")
        else:
            print(f"⚠ 图像不存在: {img_path}")

    print(f"\n✓ 成功测试了 {successful_tests}/{len(test_images)} 张真实图像")
    
    print("\n" + "=" * 60)
    print("预训练权重推理测试完成")
    print("=" * 60)
    
    return True


def compare_with_pytorch():
    """与PyTorch版本对比（如果可能）"""
    print("\n" + "=" * 40)
    print("与PyTorch版本对比")
    print("=" * 40)
    
    try:
        # 尝试加载PyTorch版本进行对比
        import torch
        sys.path.insert(0, '../nanodet-pytorch')
        
        # 这里可以添加PyTorch版本的对比代码
        print("⚠ PyTorch对比功能待实现")
        
    except Exception as e:
        print(f"⚠ 无法进行PyTorch对比: {e}")


def main():
    """主函数"""
    print("Jittor NanoDet 预训练权重推理测试")
    
    # 测试预训练推理
    success = test_pretrained_inference()
    
    if success:
        print("\n🎉 预训练权重推理测试成功！")
        print("✓ 模型架构正确")
        print("✓ ImageNet预训练权重加载成功")
        print("✓ 前向传播正常")
    else:
        print("\n❌ 预训练权重推理测试失败")
        return False
    
    # 对比测试
    compare_with_pytorch()
    
    return True


if __name__ == '__main__':
    success = main()
    sys.exit(0 if success else 1)
