nanodet-jittor/requirements.txt
# Jittor深度学习框架
jittor>=1.3.0

# 基础科学计算库
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python>=4.5.0
Pillow>=8.0.0

# 配置管理
omegaconf>=2.0.1
pyaml>=21.0.0

# 数据处理
pycocotools>=2.0.4
Cython>=0.29.0

# 可视化和日志
tensorboard>=2.8.0
tqdm>=4.62.0
termcolor>=1.1.0
tabulate>=0.8.0

# 模型导出
onnx>=1.10.0
onnx-simplifier>=0.3.0

# 其他工具
scipy>=1.7.0
scikit-image>=0.18.0
